Restructure Folder + new algorithm
- Bisection
- css folder
