+++ client/src/App.js
@@ -1,13 +1,23 @@
 import React from "react";
-import "./App.css";
+import "./css/Global/App.css";
+import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import Header from "./component/Header/Header";
 import GoldenSectionSearch from "./Optimize/GoldenSectionSearch/GoldenSectionSearch";
+import Bisection from "./Optimize/Bisection/Bisection";
 
 export default function App() {
   return (
     <div className="app-container">
-      <Header />
-      <GoldenSectionSearch />
+      <Router>
+        <Header />
+        <Routes>
+          <Route
+            path="/golden-section"
+            element={<GoldenSectionSearch />}
+          ></Route>
+          <Route path="/bisection" element={<Bisection />}></Route>
+        </Routes>
+      </Router>
     </div>
   );
 }
